# Avoid NaN-propagation in scalar rules