@sethaxen 2022-03-13
The current tests fail because
1. `zero(::NotImplemented)` throws a `NotImplementedException`
2. `zero(NoTangent())` is a `ZeroTangent()`, so this change breaks inferrability. This causes the ChainRules integration tests to fail for `copysign` and `ldexp`, for which one of the partials is `NoTangent()`