@oxinabox 2022-05-17
Sorry for the slow reply

Not yet having looked at the code, but the general idea that the zero from `@scalar_rule` should be a strong zero is correct.
And indeed it used to be a `ZeroTangent()`, but we changed it due to issues with this causes type widening.
(These issues would be resolved if we had https://github.com/JuliaLang/julia/issues/38241)


> `zero(::NotImplemented)` throws a `NotImplementedException`

This feels like it should be a `ZeroTangent()`
As it is one of the things that is in fact safe to do even if the tangent wasn't implemented.
Since we don't care about it

> `zero(NoTangent()) is a ZeroTangent()` 

This could well be changed to `zero(NoTangent()) isa NoTangent)`