@devmotion 2022-10-13
I just ran into the `sqrt` issue. Do you think you'll be able to finish and merge this PR soon, @sethaxen? Or would you like some help here?