@devmotion 2022-10-13
>> zero(::NotImplemented) throws a NotImplementedException
>
> This feels like it should be a ZeroTangent()
As it is one of the things that is in fact safe to do even if the tangent wasn't implemented.
Since we don't care about it

It seems it was changed at some point but maybe it should be reverted. It seems it causes type inference issues since now the return type can't be inferred if `NotImplemented` is involved.