@sethaxen 2022-10-15
> So it seems in the end it is not necessary to change `zero(::NotImplemented)` and `zero(::Type{<:NotImplemented})`? Maybe to be sure add tests with a `@scalar_rule` where one partial is `NotImplemented` (similar to some of the rules in SpecialFunctions) and test them with `ZeroTangent`, `NoTangent`, and `0.0` (similar to the tests for `suminv`)?

It's not necessary in the sense that we're keeping the old behavior for `NotImplemented`, The right thing to do is probably @oxinabox's suggestion in https://github.com/JuliaDiff/ChainRulesCore.jl/pull/551#issuecomment-1129121382 (making `zero(::NotImplemented) = ZeroTangent()`, but this causes inferred types of scalar rules with non-implemented (co)tangents to be type unions. See e.g. https://github.com/JuliaDiff/ChainRulesCore.jl/actions/runs/3255105837/jobs/5344073780#step:6:204. 

Or do you think that's acceptable for these rules?