@sethaxen 2022-12-29
> However, could you add more tests? For instance, involving Tangent, Thunk, and NotImplemented? To me it seems, it should be possible to run into the type inference issue for all types of partials where `typeof(partial) !== typeof(zero(partial))`.

`Tangent` type for tangents of scalar primals is not supported, so type-inferrability here is not a concern. This PR does not change the behavior for `NotImplemented` and `Thunk` (`iszero` either errors or returns `false` for both, respectively). For `NotImplemented`, type-inferrability is not a problem since an error is raised anyways. For `Thunk`, I added type-inferrability tests.

A future PR could work out how to support hard zero tangents for `NotImplemented` and `Thunk`.