@devmotion 2022-12-30
> (`iszero` either errors or returns `false` for both, respectively).

I don't follow - I was concerned about partials of these types ad for these only the type of `zero(partial)` should be relevant, and in particular we do not call `iszero` on these (I think).
With the current implementation of `NotImplemented`, `zero` will throw an error - but changing it to return `ZeroTangent()`, as proposed above, would cause type inference issues.
Maybe it's safest to replace `partial` with `zero(partial)` only if its type is Number, ZeroTangent, or NoTangent.