@sethaxen 2023-01-18
> @sethaxen can we merge this? I would like to, it seems good enough for now. We can always make follow ups

I agree that feature-wise it's good enough for a merge, but we should soon address the issues @devmotion has raised.

But before merging, it seems we fail IntegrationTests with ChainRulesTestUtils and Diffractor; I can look into the former to make sure they're not due to a problem in this PR. The failures in ChainRules are due to broken tests unexpectedly passing.