@mcabbott 2023-01-25
Not sure re Diffractor. Some chance that some 2nd order test doesn't like the extra branches?

What's the argument for `iszero(Δs_i) ? zero(∂s_i) : ∂s_i` instead of `ifelse`? The branch does not defer any calculation.

One other comment is that rules using `derivatives_given_output` will now differ, unless they re-implement this `iszero` check. Should it then be a function like `strong_times(x, y)` or is the hassle of having another name here worse than just re-doing it elsewhere (e.g. in CR?)