@devmotion 2023-01-25
> What's the argument for `iszero(Δs_i) ? zero(∂s_i) : ∂s_i` instead of `ifelse`? The branch does not defer any calculation.

IIRC the reason was https://github.com/JuliaDiff/ChainRules.jl/pull/599#discussion_r825294117.