@sethaxen 2023-01-25
> One other comment is that rules using `derivatives_given_output` will now differ, unless they re-implement this `iszero` check. Should it then be a function like `strong_times(x, y)` or is the hassle of having another name here worse than just re-doing it elsewhere (e.g. in CR?)

I don't really understand how `derivatives_given_output` is used. Do you have a suggestion for how this could be done in that function instead of during propagation?