@mcabbott 2023-01-25
I see. If there's a trade-off between pleasing 2nd order AD by avoiding branches, and a marginal speed change should anyone ever differentiate BigFloat, then we should pick AD.

`derivatives_given_output` is used e.g. here:

https://github.com/JuliaDiff/ChainRules.jl/blob/9adf759bc63432dc518ccf499d6938fc5a217113/src/rulesets/Base/mapreduce.jl#L89-L90

I wondered whether `dyₖ * conj(∂yₖ∂xᵢ)` there ought to be `ChainRulesCore. strong_times(dyₖ, conj(∂yₖ∂xᵢ))` (or perhaps `strong_dot` absorbs the `conj`) to share code. In broadcasting-like use, for simple functions, there is some chance we ought to care about SIMD here.