@sethaxen 2023-02-05
@mcabbott I added `strong_mul` and `strong_muladd` (since conjugation is performed elsewhere in the macro). Is this what you had in mind?